+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Unreleased
 
 - `all-is-cubes-mesh` library:
-    - Blocks in a `dynamic::ChunkedSpaceMesh` may now be rendered via instancing. `dynamic::DynamicMeshTypes::MINIMUM_INSTANCE_INDEX_COUNT` controls whether this is done.
+    - Blocks in a `dynamic::ChunkedSpaceMesh` may now be rendered via instancing. `dynamic::DynamicMeshTypes::MAXIMUM_MERGED_BLOCK_MESH_SIZE` controls whether this is done.
 
 ## 0.7.1 (2024-01-27)